Refactor grade handling to use FLOAT8 for Earned and Total fields in GradeResponse and related DTOs

+++ server/internal/transport/http/dto/grades_dto.go
@@ -6,8 +6,8 @@ type GradeResponse struct {
 	ID             string    `json:"id"`
 	CourseUUID     string    `json:"course_uuid"`
 	AssignmentName string    `json:"assignment_name"`
-	Earned         *int      `json:"earned"`
-	Total          *int      `json:"total"`
+	Earned         *float64  `json:"earned"`
+	Total          *float64  `json:"total"`
 	Status         string    `json:"status"`
 	PostedDate     time.Time `json:"posted_date"`
 }
@@ -16,8 +16,8 @@ type GradeMutationResponse struct {
 	ID             string    `json:"id"`
 	CourseUUID     string    `json:"course_uuid"`
 	AssignmentName string    `json:"assignment_name"`
-	Earned         *int      `json:"earned"`
-	Total          *int      `json:"total"`
+	Earned         *float64  `json:"earned"`
+	Total          *float64  `json:"total"`
 	Status         string    `json:"status"`
 	PostedDate     time.Time `json:"posted_date"`
 }
@@ -25,16 +25,16 @@ type GradeMutationResponse struct {
 type CreateGradeRequest struct {
 	CourseUUID     string    `json:"course_uuid" validate:"required,uuid"`
 	AssignmentName string    `json:"assignment_name" validate:"required"`
-	Earned         *int      `json:"earned" validate:"omitempty,gt=0"`
-	Total          *int      `json:"total" validate:"omitempty,gt=0"`
+	Earned         *float64  `json:"earned" validate:"omitempty,gt=0"`
+	Total          *float64  `json:"total" validate:"omitempty,gt=0"`
 	Status         string    `json:"status" validate:"required,oneof=GRADED UNGRADED"`
 	PostedDate     time.Time `json:"posted_date" validate:"required"`
 }
 
 type UpdateGradeRequest struct {
 	AssignmentName *string    `json:"assignment_name"`
-	Earned         *int       `json:"earned" validate:"omitempty,gt=0"`
-	Total          *int       `json:"total" validate:"omitempty,gt=0"`
+	Earned         *float64   `json:"earned" validate:"omitempty,gt=0"`
+	Total          *float64   `json:"total" validate:"omitempty,gt=0"`
 	Status         *string    `json:"status" validate:"omitempty,oneof=GRADED UNGRADED"`
 	PostedDate     *time.Time `json:"posted_date"`
 }
